// src/type/academicManagement.type.ts
export type TAcademicSemester = {
  _id: string;
  name: string;
  year: string;
  code: string;
  startMonth: string;
  endMonth: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
};

export type TAcademicFaculty = {
  _id: string;
  name: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
};
export type TAcademicDept = {
  _id: string;
  name: string;
  academicFaculty: {
    name?: string;
  };
  createdAt: string;
  updatedAt: string;
  __v: number;
};